feat(auth): implement backend infrastructure and authentication
- Set up Clerk authentication with middleware for route protection
- Configure Supabase database with appropriate schema and tables
- Add user synchronization between Clerk and Supabase
- Create Cloudinary integration for file storage
- Implement validation utilities for files and folders
- Add error handling and logging infrastructure
- Create webhook handler for Clerk events

This commit establishes the foundation for the application's backend,
ensuring proper authentication, database access, and file storage capabilities.
