+++ src/lib/supabase.ts
@@ -1,0 +1,27 @@
+import { createClient } from "@supabase/supabase-js";
+import { Database } from "@/types/supabase";
+
+// Initialize the Supabase client with service role key for server-side operations
+export const supabase = createClient<Database>(
+  process.env.NEXT_PUBLIC_SUPABASE_URL!,
+  process.env.SUPABASE_SERVICE_ROLE_KEY!,
+  {
+    auth: {
+      persistSession: false,
+    },
+  }
+);
+
+// Client-side Supabase instance with anon key
+export const createClientComponentClient = () => {
+  return createClient<Database>(
+    process.env.NEXT_PUBLIC_SUPABASE_URL!,
+    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
+    {
+      auth: {
+        autoRefreshToken: true,
+        persistSession: true,
+      },
+    }
+  );
+};
