+++ src/middleware.ts
@@ -1,0 +1,25 @@
+import { clerkMiddleware, createRouteMatcher } from "@clerk/nextjs/server";
+
+// Define public routes that don't require authentication
+const isPublicRoute = createRouteMatcher([
+  "/", // Landing page
+  "/sign-in(.*)", // Sign in pages (notice the change from /auth/sign-in)
+  "/sign-up(.*)", // Sign up pages (notice the change from /auth/sign-up)
+  "/api/webhooks/clerk", // Clerk webhooks
+]);
+
+export default clerkMiddleware(async (auth, req) => {
+  // Protect non-public routes
+  if (!isPublicRoute(req)) {
+    await auth.protect();
+  }
+});
+
+export const config = {
+  matcher: [
+    // Skip Next.js internals and all static files
+    "/((?!_next|[^?]*\\.(?:html?|css|js(?!on)|jpe?g|webp|png|gif|svg|ttf|woff2?|ico|csv|docx?|xlsx?|zip|webmanifest)).*)",
+    // Always run for API routes
+    "/(api|trpc)(.*)",
+  ],
+};
